refactor(client): fix re-exports and change TabbedView to TabbedLayout
