+++ packages/client/src/elements/index.ts
@@ -6,7 +6,6 @@ export { default as Card } from './Card';
 export { default as Copyright } from './Copyright';
 export { default as Form } from './Form';
 export { Checkbox, TextField } from './inputs';
-export { CardLayout, DefaultLayout, GlobalLayout } from './layouts';
+export { CardLayout, DefaultLayout, GlobalLayout, TabbedLayout } from './layouts';
 export { default as Link } from './Link';
 export { default as Snackbar } from './Snackbar';
-export { default as Tabs } from './TabbedView';
