+++ packages/client/src/elements/layouts/index.ts
@@ -1,3 +1,4 @@
 export { default as CardLayout } from './CardLayout';
 export { default as DefaultLayout } from './DefaultLayout';
 export { default as GlobalLayout } from './GlobalLayout';
+export { default as TabbedLayout } from './TabbedLayout';
